math_engine as separate module
